전시 생성 수정

+++ src/main/java/com/Hackathon/glow/exhibitionrate/domain/ExhibitionRate.java
@@ -27,7 +27,7 @@ public class ExhibitionRate {
 
     @NotNull
     @Min(1) @Max(5)
-    private Long rate;
+    private int rate;
 
 
     @ManyToOne
